update ScrimSectionHeader.jsx, UserProfile.jsx, MYCreatedScrims.jsx
